Fix beartype violations for multiprocessing types and jaxtyping dtypes
- Use actual classes (multiprocessing.queues.Queue, pool.Pool,
  sharedctypes.Synchronized) instead of factory functions for type hints,
  which beartype cannot resolve
- Fix flow_mag() return type: Float -> Bool for validity mask
- Fix num_val in patchgraph.py: Float -> Int (sum of bools is int64)
- Remove Queue[...] generic syntax in stream.py (unsupported in 3.11)

All 4 inference modes (vo-fast, slam-proximity, slam-classic, accurate)
now pass in both prod and dev (beartype) environments with identical
output.

+++ packages/dpvo/dpvo/patchgraph.py
@@ -125,7 +125,7 @@ class PatchGraph:
             beta=0.5,
         )
         flow_mg_sum: Float[Tensor, "fl"] = reduce(flow_mg * val, "1 (fl M) 1 1 -> fl", "sum", M=self.M).float()
-        num_val: Float[Tensor, "fl"] = reduce(val, "1 (fl M) 1 1 -> fl", "sum", M=self.M).clamp(min=1)
+        num_val: Int[Tensor, "fl"] = reduce(val, "1 (fl M) 1 1 -> fl", "sum", M=self.M).clamp(min=1)
         flow_mag_vals: Float[Tensor, "fl"] = torch.where(num_val > (self.M * 0.75), flow_mg_sum / num_val, torch.inf)
 
         mask = flow_mag_vals < self.cfg.backend_thresh
